Add length checks for dirsDepth
This change adds index checks before `dirDepths[collectionIndexDirDepth]` and `dirDepths[collectionNameDepth]` to avoid indexing beyond the end of the array. We *could* just check for the `collectionIndexDirDepth` and avoid the other two, but that would be more brittle in case the constants defined a few lines earlier were reordered. We could also just calculate `len(dirsDepth)` once as a small perf improvement.

This is related to the [suggestion](https://github.com/hyperledger/fabric/pull/1080#pullrequestreview-395879265) from @manish-sethi in #1080.